# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: vaehor API
  version: 0.0.0
...
paths:
...
  /api/admin/manual-drives:
    get:
      operationId: listAdminManualDrives
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema: {}
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Admin
      security:
        - ApiKeyAuth: []
    post:
      operationId: addAdminManualDrive
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema: {}
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema: {}
      security:
        - ApiKeyAuth: []
    delete:
      operationId: deleteAdminManualDrive
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema: {}
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema: {}
      security:
        - ApiKeyAuth: []
...
  /api/mobile/oauth-complete:
    get:
      operationId: completeMobileOAuthGet
      description: 'Finish browser OAuth: validate state, mint exchange token, redirect to vaehor://auth/callback.'
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
          explode: false
      responses:
        '302':
          description: Redirection
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Mobile
      security:
        - ApiKeyAuth: []
    post:
      operationId: completeMobileOAuthPost
      description: Redeem exchange token from the custom-scheme callback; returns bootstrap material for the shell.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileOAuthCompleteResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Mobile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileOAuthCompleteBody'
...
  /api/request-access:
    post:
      operationId: requestFolderAccess
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema: {}
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestCreateRequest'
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
    AccessRequestCreateRequest:
      type: object
      required:
        - folderId
        - folderName
      properties:
        folderId:
          type: string
        folderName:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          anyOf:
            - type: string
            - type: array
              items: {}
      description: Mirrors the legacy OpenAPI error shape used across the app.
...
    MobileOAuthCompleteBody:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    MobileOAuthCompleteResponse:
      type: object
      required:
        - cookieName
        - bootstrapToken
      properties:
        cookieName:
          type: string
        bootstrapToken:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
servers:
  - url: https://zee-index.duckdns.org
    description: Production server
    variables: {}
  - url: http://localhost:3000
    description: Development server
    variables: {}